+++ src/utils/transactions/getTransactionsHistory/helpers/getTransactionAsset.ts
@@ -1,6 +1,9 @@
 import { ITransactionListItemAsset } from 'lib/sdkDappUi';
 import { AssetType } from 'types/account.types';
-import { IconNamesEnum, TransactionServerStatusesEnum } from 'types/enums.types';
+import {
+  IconNamesEnum,
+  TransactionServerStatusesEnum
+} from 'types/enums.types';
 import { TransactionIconTypeEnum } from 'types/transaction-list-item.types';
 import { isContract } from 'utils/validation/isContract';
 import { getIsTransactionInvalidOrFailed } from './getIsTransactionInvalidOrFailed';